requirement.txt
pypdf
pymupdf
langchain-community
langchain-core
tiktoken
sentence-transformers
python-dotenv
faiss-cpu
openai
streamlit
python-dotenv
chromadb
ipykernel
langchain-huggingface
fastapi
pdfplumber
langchain-openai